Log volume consistency changes
Remove log volumes from instances that don't need them.
Use EBS for all other log volumes.
Reduce the log volume size for standard flavor from 250 to 120GB.
Reduce the logserver main volume for standard flavor from 500 to 250GB.
Increase the log volume size for pico from 1GB to 10GB.

PNDA-2340

+++ bootstrap-scripts/base.sh
@@ -13,13 +13,15 @@ apt-get update
 apt-get -y install xfsprogs
 
 # Mount the log volume, this is always xvdc
-echo "Mounting xvdc for logs"
-umount /dev/xvdc || echo 'not mounted'
-mkfs.xfs -f /dev/xvdc
-mkdir -p /var/log/panda
-sed -i "/xvdc/d" /etc/fstab
-echo "/dev/xvdc /var/log/panda auto defaults,nobootwait,comment=cloudconfig 0 2" >> /etc/fstab
-
+if [ -b /dev/xvdc ];
+then
+   echo "Mounting xvdc for logs"
+   umount /dev/xvdc || echo 'not mounted'
+   mkfs.xfs -f /dev/xvdc
+   mkdir -p /var/log/panda
+   sed -i "/xvdc/d" /etc/fstab
+   echo "/dev/xvdc /var/log/panda auto defaults,nobootwait,comment=cloudconfig 0 2" >> /etc/fstab
+fi
 # Mount the other log volumes if they exist, up to 3 more may be mounted but this list could be extended if required
 DISKS="xvdd xvde xvdf"
 DISK_IDX=0
